requirements.txt
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.0
tensorflow>=2.15.0
keras>=2.15.0
yfinance>=0.2.18
alpha-vantage>=2.3.1
newsapi-python>=0.2.6
requests>=2.31.0
matplotlib>=3.7.2
seaborn>=0.12.2
plotly>=5.15.0
ta>=0.10.2
python-dotenv>=1.0.0
schedule>=1.2.0
ccxt>=4.0.0
nltk>=3.8.1
textblob>=0.17.1
vaderSentiment>=3.3.2
xgboost>=2.0.0
lightgbm>=4.0.0
joblib>=1.3.2 